requirements.txt
python-dotenv==1.0.0
pydantic>=2.5.0
rich==13.7.0
typer==0.12.3
watchdog==3.0.0
openai>=1.54.0
google-generativeai>=0.8.0
